Move Snippet to Widget\Injector
