+++ src/EventSubscriber/WidgetsSubscriber.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Bolt\EventSubscriber;
 
-use Bolt\Snippet\RequestZone;
-use Bolt\Snippet\Target;
+use Bolt\Widget\Injector\RequestZone;
+use Bolt\Widget\Injector\Target;
 use Bolt\Widget\BoltHeaderWidget;
 use Bolt\Widget\CanonicalLinkWidget;
 use Bolt\Widget\NewsWidget;
