+++ src/Widget/Injector/HtmlInjector.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Bolt\Snippet;
+namespace Bolt\Widget\Injector;
 
 use Bolt\Common\Str;
+use Bolt\Widget\Injector\Target;
 use Bolt\Widget\ResponseAware;
 use Bolt\Widget\WidgetInterface;
 use Symfony\Component\HttpFoundation\Response;
