+++ src/Widget/Injector/QueueProcessor.php
@@ -2,14 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Bolt\Snippet;
+namespace Bolt\Widget\Injector;
 
+use Bolt\Widget\Injector\HtmlInjector;
 use Symfony\Component\HttpFoundation\Response;
 use Tightenco\Collect\Support\Collection;
 
-/**
- * Snippet queue processor.
- */
 class QueueProcessor
 {
     /** @var HtmlInjector */
